Changed aboutus to Dota
Changed About us to Dota
